// frontend/src/api/model/vaultWithNotes.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Backend API
 * OpenAPI spec version: 0.1.0
 */
import type { NoteRead } from "./noteRead";

export interface VaultWithNotes {
  created_at: string;
  id: string;
  name: string;
  notes?: NoteRead[];
  theme: string;
  updated_at: string;
}